requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rouge-score>=0.1.2
nltk>=3.8
scikit-learn>=1.1.0
tqdm>=4.64.0
python-dotenv>=0.19.0
pyyaml>=6.0